don't use envvars
